Refine README and add donation appeal
Simplified the project title in the README to be more inclusive of
multiple growth standards beyond WHO. Removed sections detailing work in
progress and specific usage instructions, consolidating the information
with an example section. Introduced a call-to-action for community
support with links to GitHub Sponsors and Ko-Fi, aiming to encourage
donations to sustain and advance project development. This change seeks
to engage the community more effectively and secure resources for
ongoing maintenance and enhancement.